+++ projects/gloo/pkg/plugins/advanced_http/plugin.go
@@ -1,4 +1,4 @@
-package http_path
+package advanced_http
 
 import (
 	envoy_config_cluster_v3 "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
@@ -15,8 +15,8 @@ var (
 )
 
 const (
-	ErrEnterpriseOnly = "Could not load http_path plugin to configure custom paths/endpoint - this is an Enterprise feature"
-	ExtensionName     = "http_path"
+	ErrEnterpriseOnly = "Could not load advanced_http plugin to configure custom paths/methods per endpoint, or complex health check response parsing - this is an Enterprise feature"
+	ExtensionName     = "advanced_http"
 )
 
 type plugin struct{}
@@ -40,6 +40,15 @@ func (p *plugin) Init(params plugins.InitParams) error {
 func (p *plugin) ProcessUpstream(params plugins.Params, in *v1.Upstream, out *envoy_config_cluster_v3.Cluster) error {
 	for _, host := range in.GetStatic().GetHosts() {
 		if host.GetHealthCheckConfig().GetPath() != "" {
+			return eris.New(ErrEnterpriseOnly)
+		}
+		if host.GetHealthCheckConfig().GetMethod() != "" {
+			return eris.New(ErrEnterpriseOnly)
+		}
+	}
+
+	for _, hc := range in.GetHealthChecks() {
+		if len(hc.GetHttpHealthCheck().GetResponseAssertions().GetResponseMatchers()) != 0 {
 			return eris.New(ErrEnterpriseOnly)
 		}
 	}
